chore: Moves Dropdown to the components folder (#14126)
